refactor: remove redundant base class and innner class merged
