+++ src/main/java/mrriegel/storagenetwork/gui/GuiContainerBase.java
@@ -1,10 +1,7 @@
 package mrriegel.storagenetwork.gui;
 
-import javax.annotation.Nonnull;
-import mrriegel.storagenetwork.util.UtilInventory;
+import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.gui.inventory.GuiContainer;
-import net.minecraft.client.renderer.GlStateManager;
-import net.minecraft.client.renderer.RenderHelper;
 import net.minecraft.inventory.Container;
 import net.minecraft.item.ItemStack;
 
@@ -14,66 +11,22 @@ public abstract class GuiContainerBase extends GuiContainer {
     super(inventorySlotsIn);
   }
 
-  /**
-   * used as the MAIN grid in the network item display
-   * 
-   * also as ghost/filter items in the cable filter slots
-   * 
-   * @author
-   *
-   */
-  public class ItemSlotNetwork extends AbstractSlot {
+  @Override
+  public void drawGradientRect(int left, int top, int right, int bottom, int startColor, int endColor) {
+    super.drawGradientRect(left, top, right, bottom, startColor, endColor);
+  }
 
-    public ItemStack stack;
+  public FontRenderer getFont() {
+    return this.fontRenderer;
+  }
 
-    public ItemSlotNetwork(GuiContainerBase parent, @Nonnull ItemStack stack, int x, int y, int size, int guiLeft, int guiTop, boolean number, boolean square, boolean smallFont, boolean toolTip) {
-      super(parent, x, y, size, guiLeft, guiTop, number, square, smallFont, toolTip);
-      this.stack = stack;
-    }
+  @Override
+  public boolean isPointInRegion(int rectX, int rectY, int rectWidth, int rectHeight, int pointX, int pointY) {
+    return super.isPointInRegion(rectX, rectY, rectWidth, rectHeight, pointX, pointY);
+  }
 
-    @Override
-    public boolean isMouseOverSlot(int mouseX, int mouseY) {
-      return parent.isPointInRegion(x - guiLeft, y - guiTop, 16, 16, mouseX, mouseY);
-    }
-
-    @Override
-    public void drawSlot(int mx, int my) {
-      GlStateManager.pushMatrix();
-      if (!stack.isEmpty()) {
-        RenderHelper.enableGUIStandardItemLighting();
-        mc.getRenderItem().renderItemAndEffectIntoGUI(stack, x, y);
-        String amount = UtilInventory.formatLargeNumber(size);
-        if (number) {
-          if (smallFont) {
-            GlStateManager.pushMatrix();
-            GlStateManager.scale(.5f, .5f, .5f);
-            mc.getRenderItem().renderItemOverlayIntoGUI(parent.fontRenderer, stack, x * 2 + 16, y * 2 + 16, amount);
-            GlStateManager.popMatrix();
-          }
-          else {
-            mc.getRenderItem().renderItemOverlayIntoGUI(parent.fontRenderer, stack, x, y, amount);
-          }
-        }
-      }
-      if (square && this.isMouseOverSlot(mx, my)) {
-        GlStateManager.disableLighting();
-        GlStateManager.disableDepth();
-        int j1 = x;
-        int k1 = y;
-        GlStateManager.colorMask(true, true, true, false);
-        drawGradientRect(j1, k1, j1 + 16, k1 + 16, -2130706433, -2130706433);
-        GlStateManager.colorMask(true, true, true, true);
-        GlStateManager.enableLighting();
-        GlStateManager.enableDepth();
-      }
-      GlStateManager.popMatrix();
-    }
-
-    @Override
-    public void drawTooltip(int mx, int my) {
-      if (toolTip && this.isMouseOverSlot(mx, my) && !stack.isEmpty()) {
-        renderToolTip(stack, mx, my);
-      }
-    }
+  @Override
+  public void renderToolTip(ItemStack stack, int x, int y) {
+    super.renderToolTip(stack, x, y);
   }
 }
